Refactor components to handle new store structure
