+++ client/package.json
@@ -30,6 +30,7 @@
     "@types/react-redux": "^7.1.16",
     "@types/react-router-dom": "^5.1.7",
     "@types/react-slick": "^0.23.4",
+    "@types/uuid": "^8.3.0",
     "date-fns": "^2.19.0",
     "history": "4.10.1",
     "konva": "^7.2.5",
@@ -45,6 +46,7 @@
     "redux-thunk": "^2.3.0",
     "typescript": "^4.1.2",
     "use-image": "^1.0.7",
+    "uuid": "^8.3.2",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
