+++ client/src/components/compose/Authorizers.tsx
@@ -2,95 +2,23 @@ import React from 'react';
 
 // Libraries
 import {makeStyles, Container, Typography} from '@material-ui/core';
-import {FontAwesomeIcon} from '@fortawesome/react-fontawesome';
-import {faTrashAlt} from '@fortawesome/free-solid-svg-icons';
+import {v4 as uuidv4} from 'uuid';
 
 // Context hooks
 import {useCompose} from '../../store/contexts';
-import {AuthorizerType} from '../../store/action-types';
-import {COMPOSE} from '../../store/types';
+import {AuthorizerType, COMPOSE} from '../../store/action-types';
 
 // Components
-import {CustomButton, CustomTextInput} from '../shared';
-
-// Hooks
-import {useInput} from '../../hooks';
-
-function Authorizer({
-  item,
-  removeAuthorizer,
-}: {
-  item: AuthorizerType;
-  removeAuthorizer: (item: string) => void;
-}): JSX.Element {
-  const [authorizer, setAuthorizer] = useInput();
-  const [message, setMessage] = useInput();
-
-  const classes = useStyles();
-  return (
-    <div className={classes.authorizerContainer}>
-      <div className={classes.authorizerRow1}>
-        <CustomTextInput
-          value={authorizer}
-          onChange={setAuthorizer}
-          select
-          required
-          variant="outlined"
-          label="Authorizer"
-          className={classes.authorizerInput1}
-          options={[
-            {
-              label: 'Prof. Abel Mathew',
-              value: 'Prof. Abel Mathew',
-            },
-            {
-              label: 'Prof. Harish R.',
-              value: 'Prof. Harish R.',
-            },
-            {
-              label: 'Prof. Ritesh Patil',
-              value: 'Prof. Ritesh Patil',
-            },
-          ]}
-        />
-        <FontAwesomeIcon
-          className={classes.deleteAuthorizer}
-          size="2x"
-          icon={faTrashAlt}
-          onClick={() => removeAuthorizer(item.id)}
-        />
-      </div>
-
-      <CustomTextInput
-        value={message}
-        onChange={setMessage}
-        multiline
-        required
-        variant="outlined"
-        label="Message to the Authorizer"
-        className={classes.authorizerInput2}
-      />
-    </div>
-  );
-}
+import {CustomButton} from '../shared';
+import Authorizer from './Authorizer';
 
 function Authorizers(): JSX.Element {
-  // const [authorizers, setAuthorizers] = React.useState<string[]>([]);
-
   const [state, dispatch] = useCompose();
-
-  // const addAuthorizer = () =>
-  //   setAuthorizers(current => [...current, `authorizer${Math.random() * 10}`]);
-
-  // const removeAuthorizer = (item: string): void =>
-  //   setAuthorizers(current =>
-  //     current.filter(authorizer => authorizer !== item),
-  //   );
 
   const addAuthorizer = () =>
     dispatch({
       type: COMPOSE.ADD_NEW_AUTHORIZER,
-      payload: {id: Math.random() * 10},
+      payload: {id: uuidv4()},
     });
 
   const removeAuthorizer = (id: string) =>
@@ -116,33 +44,24 @@ function Authorizers(): JSX.Element {
           onClick={addAuthorizer}
         />
       </div>
+
       {state.authorizerDetails.length > 0 &&
-        state.authorizerDetails.map(authorizer => (
-          <Authorizer
-            key={authorizer.id}
-            item={authorizer}
-            removeAuthorizer={removeAuthorizer}
-          />
-        ))}
-      {/* {authorizers.length > 0 &&
-        authorizers
-          // .reverse()
-          .map(
-            (authorizer: string): JSX.Element => (
-              <Authorizer
-                key={authorizer}
-                item={authorizer}
-                removeAuthorizer={removeAuthorizer}
-              />
-            ),
-          )} */}
+        state.authorizerDetails.map(
+          (authorizer: AuthorizerType): JSX.Element => (
+            <Authorizer
+              key={authorizer.id}
+              item={authorizer}
+              removeAuthorizer={removeAuthorizer}
+            />
+          ),
+        )}
     </Container>
   );
 }
 
 export default Authorizers;
 
-const useStyles = makeStyles(theme => ({
+const useStyles = makeStyles(() => ({
   root: {
     width: '100%',
     height: '100%',
@@ -171,32 +90,5 @@ const useStyles = makeStyles(theme => ({
     width: '45px',
     height: '45px',
     minWidth: '45px',
-  },
-  authorizerContainer: {
-    padding: 0,
-    marginTop: '20px',
-    borderBottom: '1px solid rgba(0,0,0,0.23)',
-    width: '100%',
-  },
-  authorizerInput1: {
-    margin: '10px 0px',
-    width: '80%',
-  },
-  authorizerInput2: {
-    margin: '10px 0px',
-  },
-  authorizerRow1: {
-    display: 'flex',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-  },
-  deleteAuthorizer: {
-    width: '15%',
-    marginRight: '10px',
-    // @ts-ignore
-    color: theme.palette.accent.red,
-    '&:hover': {
-      cursor: 'pointer',
-    },
   },
 }));
